Move ByID functions to separate file.
